RegistryHolder Registered response message classes should be private
