+++ runtime/src/main/resources/local.conf
@@ -24,8 +24,8 @@ akka {
       "com.spbsu.flamestream.runtime.master.acker.api.registry.RegisterFront" = kryo
       "com.spbsu.flamestream.runtime.master.acker.api.registry.RegisterFrontFromTime" = kryo
       "com.spbsu.flamestream.runtime.master.acker.api.registry.FrontTicket" = kryo
-      "com.spbsu.flamestream.runtime.master.acker.api.registry.RegisteredNewFront" = kryo
-      "com.spbsu.flamestream.runtime.master.acker.api.registry.RegisteredAlreadyRegisteredFront" = kryo
+      "com.spbsu.flamestream.runtime.master.acker.RegistryHolder$NewFrontRegisterer$Registered" = kryo
+      "com.spbsu.flamestream.runtime.master.acker.RegistryHolder$AlreadyRegisteredFrontRegisterer$Registered" = kryo
       "com.spbsu.flamestream.runtime.graph.api.AddressedItem" = kryo
       "com.spbsu.flamestream.runtime.master.acker.api.registry.UnregisterFront" = kryo
       "com.spbsu.flamestream.runtime.master.acker.api.commit.GimmeLastCommit" = kryo
